move function to new module
